+++ src/open_prime_hunters_rando/entities/entity_data.py
